requirements_flask.txt
# Flask Web App Requirements
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1

# AWS SDK for Polly
boto3>=1.26.0

# E-Commerce Authentication
flask-jwt-extended==4.6.0
bcrypt==4.1.2

# Existing requirements
streamlit>=1.28.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain>=0.1.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
streamlit-js-eval>=0.1.5
geopy>=2.3.0
requests>=2.31.0
Pillow>=10.0.0
pymysql
mcp
